requirements.txt
Flask
mediapipe
numpy
opencv-python
Pillow